tools/dhcpns/pyproject.toml
[tool.poetry]
name = "dhcpns"
version = "0.1.0"
description = ""
authors = ["Ole Mathias Heggem <olemathias.aa.heggem@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.28.2"
pynetbox = "^7.0.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
